Update totem and att restrictions
