+++ server/game/cards/06-TLS/FireOfNanahbozho.js
@@ -3,13 +3,26 @@ const SpellCard = require('../../spellcard.js');
 
 class FireOfNanahbozho extends SpellCard {
     setupCardAbilities(ability) {
+        this.attachmentRestriction({ type: 'deed'});
+        this.whileAttached({
+            effect: ability.effects.addKeyword('holy ground')
+        });
+        this.persistentEffect({
+            targetController: 'any',
+            condition: () => true,
+            match: card => card.location === 'play area' && 
+                card.getType() === 'dude' &&
+                card.gamelocation === this.gamelocation && 
+                card.hasKeyword('shaman'),
+            effect: ability.effects.modifySkillRating('shaman', 2)
+        });
         this.spellAction({
             title: 'Fire of Nanahbozho',
             playType: ['noon', 'shootout'],
             cost: ability.costs.bootSelf(),
             target: {
                 activePromptTitle: 'Select a dude',
-                cardCondition: { location: 'play area', controller: 'current' },
+                cardCondition: { location: 'play area', controller: 'current', condition: card => card.gamelocation === this.gamelocation },
                 cardType: ['dude'],
                 gameAction: 'unboot'
             },
