+++ server/game/goodscard.js
@@ -1,27 +1,6 @@
-const AbilityDsl = require('./abilitydsl.js');
-const DrawCard = require('./drawcard.js');
+const HeartsCard = require('./heartscard.js');
 
-class GoodsCard extends DrawCard {
-    constructor(owner, cardData, useMeleeEffect) {
-        super(owner, cardData);
-        this.traded = false;
-        this.canTrade = true;
-        this.resetHandler = () => this.reset();
-
-        if(this.bullets) {
-            this.whileAttached({
-                condition: () => !useMeleeEffect || !this.meleeWeaponCondition(),
-                effect: AbilityDsl.effects.dynamicBullets(() => this.bullets)
-            });
-        }
-        if(this.influence) {
-            this.whileAttached({
-                condition: () => true,
-                effect: AbilityDsl.effects.dynamicInfluence(() => this.influence)
-            });
-        }
-    }
-
+class GoodsCard extends HeartsCard {
     canAttach(player, card, playingType) {
         if(!super.canAttach(player, card)) {
             return false;
@@ -31,76 +10,13 @@ class GoodsCard extends DrawCard {
             return false;
         }
 
-        if(card.getType() === 'dude') {
-            if(this.hasKeyword('weapon') && !card.canAttachWeapon(this)) {
-                return false;
-            } 
-            if(this.hasKeyword('horse') && !card.canAttachHorse(this)) {
-                return false;
-            }
-            if(this.hasKeyword('attire') && !card.canAttachAttire(this)) {
-                return false;
-            }
-
-            return true;
-        }
-
         if(card.getType() === 'deed') {
             if(!this.hasKeyword('improvement')) {
                 return false;
             }
-            return true;
         }        
 
-        return false;
-    }
-
-    meleeWeaponCondition() {
-        return this.game.shootout && this.game.shootout.getParticipants().some(dude => {
-            if(dude.controller === this.controller) {
-                return false;
-            }
-            let nonMeleeUnbootedWeapon = dude.attachments.filter(att => att.hasKeyword('weapon') && !att.hasKeyword('melee') && !att.booted);
-            return nonMeleeUnbootedWeapon && nonMeleeUnbootedWeapon.length > 0;
-        });
-    }
-
-    canBeTraded() {
-        return this.canTrade;
-    }
-
-    wasTraded() {
-        return this.traded;
-    }
-
-    entersPlay() {
-        super.entersPlay();
-        this.game.on('onRoundEnded', this.resetHandler);
-    }
-
-    leavesPlay() {
-        super.leavesPlay();
-        this.game.removeListener('onRoundEnded', this.resetHandler);
-    }
-
-    reset() {
-        this.traded = false;
-    }
-
-    isHex() {
-        return this.hasKeyword('Hex');
-    }
-
-    isMiracle() {
-        return this.hasKeyword('Miracle');
-    }
-
-    isSpirit() {
-        return this.hasKeyword('Spirit');
-    }
-
-    isTotem() {
-        return this.hasKeyword('Totem');
+        return true;
     }
 }
